Handle recursive directory traversal
Use a simple stack to store and subsequently traverse any directories we
find during processing.
Currently the stack has finite memory and if its limit is reached the
program will crash. This is fine for now, but we'll want to decide
whether to go with more dynamic storage, or perhaps to trigger a batch
process to clear the stack before continuing.

Rearrange how we handle lack of trailing slashes for directory paths.

Make sure to clean up our heap allocations and avoid memory leaks.

Also:
- Move to using a source directory
- Separate out some utility code
- Update Makefile to handle multiple source files