Improve mp4 detection
Previously, we risked processing uninitialised data.
If a file being tested was less than the size of the header under test,
the struct would include garbage bytes.

Worse, if a file was empty, the garbage data present could contain data
from a previous read, where if a successful match was made, would lead
to a false positive.