Rewrite line reader
Our previous method for reading input was brittle. While we wanted to
move to trusting PATH_MAX, that is apparently ill-advised.
See: https://insanecoding.blogspot.com/2007/11/pathmax-simply-isnt.html

Instead we opted for an adaptive read, where we start with a default
length, and grow only if necessary.
Once the buffer is enlarged, the new allocation is used for future
reads.

Also:
- Update existing code to make proper use of the new function

TODO:
- Implement some reasonable maximum buffer length, as currently we keep
  reading indefinitely until a newline char is found