Retain info on tracked directories
First pass at keeping data on each of the directories a user has set to
be tracked.
Liable to change significantly over time.
For now, we just store the path of a tracked directory, along with the
videos specific to that directory.

This should enable us to prevent tracking the same directories more than
once, as well as prevent and warn against trying to track parent or
child directories of those already being tracked.