Improve Makefile
When using vim, and likely similar editors, it's very handy to have a
tags file to aid in navigating the codebase.

Less necessary, but arguably nice to have is some highlighting of custom
types. This has been set up specifically for vim and the author's
personal config.
As such, it's not ideal that it is in the shared codebase.

TODO:
- Decide on a way to separate out personal config from the shared repo